producer/requirements.txt
memphis_py==0.2.7
